Backend/pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtpd>=1.4.6",
    "atpublic>=5.1",
    "black>=25.1.0",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "django-stubs-ext>=5.1.3",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-stubs>=3.15.3",
    "mypy>=1.15.0",
    "selenium>=4.31.0",
]
